requirements.txt
Flask
transformers
torch
python-docx
PyMuPDF
gunicorn
python-dotenv
requests